requirements.txt
mysql-connector-python>=8.0.0
pandas>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0